requirements.txt
pip
autopep8
Flask==3.0.3
flask_cors
gunicorn==22.0.0
Werkzeug==3.0.6
python-dotenv==1.0.1
RapidFuzz==3.12.2
google-genai